Update person and book display functionality
